requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# LeagueOfChance/main.py: 8
beautifulsoup4 == 4.8.2

# LeagueOfChance/main.py: 1,2
discord.py == 1.7.3

# LeagueOfChance/main.py: 5
python-dotenv == 0.19.2

# LeagueOfChance/main.py: 9
requests == 2.22.0
